Added user sign up/ sign using firebase.
App component: creation of AuthContext to hold user authentication information, added autentication routes and layout route for  authentication.

AuthRequired component: used in the layout route for the routes that need authentication.

SignUp component: new component that allows the user to create an account.

SignIn component: new component that allows the user to sign in to existing account.